WIP: trying to isolate constexpr failures
