+++ include/universal/internal/bitblock/exceptions.hpp
@@ -8,9 +8,11 @@
 #include <stdexcept>
 #include <string>
 
-#ifndef BITBLOCK_THROW_ARITHMETIC_EXCEPTION
-#define BITBLOCK_THROW_ARITHMETIC_EXCEPTION 0
-#endif
+// this is a supporting class, and its exception behavior needs to be configure by the calling environment.
+// this means that the value of BITBLOCK_THROW_ARITHMETIC_EXCEPTION must be set externally
+// #define BITBLOCK_THROW_ARITHMETIC_EXCEPTION exception-config-of-calling-class
+// TODO: do we want to support the use-case where two seperate classes that use bitblock<> 
+// but desire different exception behavior?
 
 namespace sw::universal {
